Clean local state in CacheStep
